Fixed ippeveprinter call in "list" to work with arbitrary backends
As ippprint takes the backend via env variable DEVIE_URI and runs the backend by itself, we set the variable and do not supply the "-D" option to ippeveprinter, as ippeveprinter does not support the actual CUPS backends.

We also supply paths for the CUPS filters, backends, and MIME rules to make ippprint correctly working.