Correctly handle detected but unsupported printers
Sometimes a printer gets detected and a device URI for it found with the available backends but now PPD for it is available in this package, meaning that the printer is not supported by this package. Resgister also these printers to avoid a repeated PPD search every 10 seconds. Also write debug messages correctly into the log file.