Use unique service names for each IPP printer emulation
The service name contains the backend name, so that if the printer is accessible through more than one backend, for each backend there will be a working IPP printer emulation.

The service name contains also the printer's serial number, to allow several printers of the same model connected to one machine.

Optionally the server can be started with the environment variable PACKAGENAME set to add an identification of the Printer Application to the service name, so that if there is more than one Printer Application supporting a printer one can easily choose.